infra/tests/requirements.txt
# Staging Test Requirements
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0
pyyaml>=6.0
